refactor: consolidate ub-moji components into dataset directory
The components/ub-moji/ and components/dataset/ directories both
contained dataset-related components with no meaningful distinction.
Merge them under dataset/ and remove the ub-moji/ directory.

Update all import paths in pages and test files accordingly.
